+++ quint/test/importResolver.test.ts
@@ -20,9 +20,11 @@ describe('resolveImports', () => {
     ],
   })
 
-  const tables: DefinitionsByModule = new Map<string, DefinitionsByName>([
-    ['wrapper', newTable({})], ['test_module', table],
-  ])
+  function tables(): DefinitionsByModule {
+    return new Map<string, DefinitionsByName>([
+      ['wrapper', newTable({})], ['test_module', table],
+    ])
+  }
 
   describe('existing modules', () => {
     it('imports named definitions', () => {
@@ -30,7 +32,7 @@ describe('resolveImports', () => {
         'import test_module.a',
       ])
 
-      const [errors, definitions] = resolveImports(quintModule, tables)
+      const [errors, definitions] = resolveImports(quintModule, tables())
 
       assert.isEmpty(errors)
       assert.deepInclude([...definitions.valueDefinitions.keys()], 'a')
@@ -42,7 +44,7 @@ describe('resolveImports', () => {
         'import test_module.*',
       ])
 
-      const [errors, definitions] = resolveImports(quintModule, tables)
+      const [errors, definitions] = resolveImports(quintModule, tables())
 
       assert.isEmpty(errors)
       assert.includeDeepMembers([...definitions.valueDefinitions.keys()], ['a', 'b'])
@@ -54,7 +56,7 @@ describe('resolveImports', () => {
         'import test_module(c1 = 3, c2 = 4) as test_module_instance',
       ])
 
-      const [errors, definitions] = resolveImports(quintModule, tables)
+      const [errors, definitions] = resolveImports(quintModule, tables())
 
 
       assert.isEmpty(errors)
@@ -71,13 +73,76 @@ describe('resolveImports', () => {
       const quintModule = buildModuleWithDefs([
         'import wrapper.*',
         'import wrapper(c1 = 1) as w',
+        'export wrapper.*',
       ])
 
-      const [errors, _definitions] = resolveImports(quintModule, tables)
+      const [errors, _definitions] = resolveImports(quintModule, tables())
 
       assert.sameDeepMembers([...errors.entries()], [
         [1n, { code: 'QNT407', message: 'Cannot import wrapper inside wrapper', data: {} }],
         [4n, { code: 'QNT407', message: 'Cannot instantiate wrapper inside wrapper', data: {} }],
+        [5n, { code: 'QNT407', message: 'Cannot export wrapper inside wrapper', data: {} }],
+      ])
+    })
+
+    it('exports all definitions', () => {
+      const quintModule = buildModuleWithDefs([
+        'export test_module.*',
+      ])
+
+      const [errors, definitions] = resolveImports(quintModule, tables())
+
+      assert.isEmpty(errors)
+      assert.deepEqual(definitions.valueDefinitions.get('a'), [
+        { kind: 'def', identifier: 'a', reference: 1n, scope: undefined },
+      ])
+      assert.deepEqual(definitions.typeDefinitions.get('T'), [{ identifier: 'T', reference: 1n, scope: undefined }])
+    })
+
+    it('exports previously imported definitions', () => {
+      const quintModule = buildModuleWithDefs([
+        'import test_module.*',
+        'export test_module.*',
+      ])
+
+      const [errors, definitions] = resolveImports(quintModule, tables())
+
+      assert.isEmpty(errors)
+      assert.deepEqual(definitions.valueDefinitions.get('a'), [
+        { kind: 'def', identifier: 'a', reference: 1n, scope: undefined },
+      ])
+      assert.deepEqual(definitions.typeDefinitions.get('T'), [{ identifier: 'T', reference: 1n, scope: undefined }])
+    })
+
+    it('exports specific definitions', () => {
+      const quintModule = buildModuleWithDefs([
+        'import test_module.*',
+        'export test_module.a',
+      ])
+
+      const [errors, definitions] = resolveImports(quintModule, tables())
+
+      assert.isEmpty(errors)
+      assert.deepEqual(definitions.valueDefinitions.get('a'), [
+        { kind: 'def', identifier: 'a', reference: 1n, scope: undefined },
+      ])
+      assert.deepEqual(definitions.typeDefinitions.get('T'), [{ identifier: 'T', reference: 1n, scope: 3n }])
+    })
+
+    it('exports definitions with qualifier', () => {
+      const quintModule = buildModuleWithDefs([
+        'import test_module.*',
+        'export test_module as my_export',
+      ])
+
+      const [errors, definitions] = resolveImports(quintModule, tables())
+
+      assert.isEmpty(errors)
+      assert.deepEqual(definitions.valueDefinitions.get('a'), [
+        { kind: 'def', identifier: 'a', reference: 1n, scope: 3n },
+      ])
+      assert.deepEqual(definitions.valueDefinitions.get('my_export::a'), [
+        { kind: 'def', identifier: 'my_export::a', reference: 1n, scope: undefined },
       ])
     })
   })
@@ -88,7 +153,7 @@ describe('resolveImports', () => {
         'import unexisting_module.*',
       ])
 
-      const [errors, _definitions] = resolveImports(quintModule, tables)
+      const [errors, _definitions] = resolveImports(quintModule, tables())
 
       assert.sameDeepMembers([...errors.entries()], [
         [1n, { code: 'QNT404', message: 'Module unexisting_module not found', data: {} }],
@@ -100,10 +165,22 @@ describe('resolveImports', () => {
         'import unexisting_module(c1 = c1, c2 = c2) as test_module_instanc',
       ])
 
-      const [errors, _definitions] = resolveImports(quintModule, tables)
+      const [errors, _definitions] = resolveImports(quintModule, tables())
 
       assert.sameDeepMembers([...errors.entries()], [
         [5n, { code: 'QNT404', message: 'Module unexisting_module not found', data: {} }],
+      ])
+    })
+
+    it('fails exporting', () => {
+      const quintModule = buildModuleWithDefs([
+        'export unexisting_module.*',
+      ])
+
+      const [errors, _definitions] = resolveImports(quintModule, tables())
+
+      assert.sameDeepMembers([...errors.entries()], [
+        [1n, { code: 'QNT404', message: 'Module unexisting_module not found', data: {} }],
       ])
     })
   })