+++ CHANGELOG.md
@@ -9,11 +9,41 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Modules and definitions can now be exported (#771)
+
+### Changed
+
+- The ITF export adds output similar to Apalache (#780)
+
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## v0.9.1 -- 2023-04-04
+
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+
+- `test` command now exits with non-zero code on test failures (#772)
+
+### Security
+
+## v0.9.0 -- 2023-04-03
+
+### Added
+
 - detailed run output on `run --verbosity=3` (#748)
 - detailed run output on `test --verbosity=3` (#755)
 - detailed run output in REPL on `.verbosity=3` (#764)
 
 ### Changed
+
+- using a controllable pRNG (#767)
+
 ### Deprecated
 ### Removed
 
@@ -21,7 +51,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Effects no longer break when applying constant operators (#759)
+- effects no longer break when applying constant operators (#759)
+- number highlighting in vim (#765)
+- the simulator supports really big numbers, e.g., 256-bit (#767)
 
 ### Security